chore: move servers struct to own file
